feat(survey): renders results page after clicking submit
The Results component handles the calculations based on the form data submitted from the survey.

Calculations:
All raw scores are out of 40, and then converted to the corresponding percentage and displayed to the user. Each personality trait - O, C, E, A, N - has a corresponding starting value. Questions are weight either negative or positive determining whether the value from the form is subtracted or added to the total raw score.

After the raw score is computed, it is designated low, moderate, or high and corresponding explanation text is displayed to the user.