feat(backend): create PersonalityResults Table and Users Table, make a seed script, and call api endpoints after personality scores are calculated
The backend uses Prisma as an ORM and SQLite as the database.

The following models were added to the Prisma schema:
model User {
  id: Int  (Primary)
  firstName: String?
  lastName: String?
  age: Int?
  zipCode: String?
  race: String?
}
only the id field is required, but that is auto-generated by Prisma

model PersonalityResult {
  id: Int (Primary)
  user: User
  userId: Int
  OpennessScore: Int
  ConscientiousnessScore: Int
  ExtraversionScore: Int
  AgreeablenessScore: Int
  NeuroticismScore: Int
}

For the User Table, a create single user endpoint was created.
For the PersonalityResult Table, a create single result endpoint and get all results endpoint was created.

In this commit, a single User and single PersonalityResult is created inside the Results component after scores are calculated.

Also made a seed script to seed database with fake data. Just need to call `npm run seed`